# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFSO API",
        "description": "API for TFSO application",
        "contact": {
            "name": "Burak",
            "url": "https://burakkaynar.com",
            "email": "burak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Giriş yapmış kullanıcının tüm favorilerini listeler",
                "operationId": "933a74d1899fde8e6df8d9fd0b9d318e",
                "responses": {
                    "200": {
                        "description": "Favorilerin listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tracks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Track"
                                            }
                                        },
                                        "sets": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Set"
                                            }
                                        },
                                        "radios": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Radio"
                                            }
                                        },
                                        "djs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/tracks": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Giriş yapmış kullanıcının favori şarkılarını listeler",
                "operationId": "d9bd32426e483299af3f596b33be84b2",
                "responses": {
                    "200": {
                        "description": "Favori şarkıların listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Track"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/sets": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Giriş yapmış kullanıcının favori setlerini listeler",
                "operationId": "8c660ad7e9f1e3b83eda2165b309f124",
                "responses": {
                    "200": {
                        "description": "Favori setlerin listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Set"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/radios": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Giriş yapmış kullanıcının favori radyolarını listeler",
                "operationId": "5b2602d58755fea60c0a0b681716f007",
                "responses": {
                    "200": {
                        "description": "Favori radyoların listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Radio"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/djs": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Giriş yapmış kullanıcının favori DJ'lerini listeler",
                "operationId": "4369a2ac271b847a70a28ff7d8eafbb6",
                "responses": {
                    "200": {
                        "description": "Favori DJ'lerin listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tracks/{track}/toggle-favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Bir şarkıyı favorilere ekler veya çıkarır",
                "operationId": "3dea4709cbc0348799b0fce8bc2445c8",
                "parameters": [
                    {
                        "name": "track",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favori durumu güncellendi"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "404": {
                        "description": "Şarkı bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sets/{set}/toggle-favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Bir seti favorilere ekler veya çıkarır",
                "operationId": "5feae6898b8ef58614fafa67c6194cff",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favori durumu güncellendi"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "404": {
                        "description": "Set bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/radios/{radio}/toggle-favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Bir radyoyu favorilere ekler veya çıkarır",
                "operationId": "ec6383dae6b792fec06af5fcfbbebaf5",
                "parameters": [
                    {
                        "name": "radio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favori durumu güncellendi"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "404": {
                        "description": "Radyo bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/djs/{dj}/toggle-favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Bir DJ'i favorilere ekler veya çıkarır",
                "operationId": "a44bc7ac773bd8d4d043699990a1e2f3",
                "parameters": [
                    {
                        "name": "dj",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favori durumu güncellendi"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "404": {
                        "description": "DJ bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model representing a user in the system",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the user",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Password for the user account",
                        "type": "string",
                        "format": "password"
                    },
                    "profile_photo": {
                        "description": "URL of the user's profile photo",
                        "type": "string"
                    },
                    "bio": {
                        "description": "Short biography of the user",
                        "type": "string"
                    },
                    "instagram": {
                        "description": "Instagram handle of the user",
                        "type": "string"
                    },
                    "twitter": {
                        "description": "Twitter handle of the user",
                        "type": "string"
                    },
                    "facebook": {
                        "description": "Facebook profile URL of the user",
                        "type": "string"
                    },
                    "tiktok": {
                        "description": "TikTok handle of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Radio": {
                "title": "Radio",
                "description": "Radio station model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "type": "string"
                    },
                    "stream_url": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "logo_url": {
                        "type": "string",
                        "readOnly": "true",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Set": {
                "title": "Set",
                "description": "Model representing a set of tracks",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the set",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "ID of the user who created the set",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the set",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the set",
                        "type": "string"
                    },
                    "cover_image": {
                        "description": "URL of the cover image",
                        "type": "string"
                    },
                    "audio_file": {
                        "description": "Path to the audio file",
                        "type": "string"
                    },
                    "is_premium": {
                        "description": "Indicates if the set is premium",
                        "type": "boolean"
                    },
                    "iap_product_id": {
                        "description": "In-app purchase product ID for premium access",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp of the set",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the set",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Track": {
                "title": "Track",
                "description": "Track model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "audio_file": {
                        "type": "string"
                    },
                    "image_file": {
                        "type": "string"
                    },
                    "is_premium": {
                        "type": "boolean"
                    },
                    "iap_product_id": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "audio_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "image_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "audio_file_name": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    },
...
}